+++ app/assets/stylesheets/pages/_dashboard.scss
@@ -191,9 +191,9 @@ html {
 .item.animate {
   z-index: 999;
   width: 80%;
-  position: fixed;
+ /*  position: fixed;
   left: 0;
-  right: 0;
+  right: 0; */
   max-height: 98%;
   overflow-y: scroll;
   /* margin-left: auto; */
